Cookie expiry & script js
